Move tests for cards to common
