+++ src/tests/unit/tests/DetailsView/components/test-view-container.test.tsx
@@ -7,7 +7,7 @@ import { It, Mock, MockBehavior } from 'typemoq';
 import { VisualizationConfiguration } from '../../../../../common/configs/visualization-configuration';
 import { VisualizationConfigurationFactory } from '../../../../../common/configs/visualization-configuration-factory';
 import { TestViewContainer, TestViewContainerProps } from '../../../../../DetailsView/components/test-view-container';
-import { exampleUnifiedStatusResults } from './cards/sample-view-model-data';
+import { exampleUnifiedStatusResults } from '../../common/components/cards/sample-view-model-data';
 
 describe('TestViewContainerTest', () => {
     it('should not return the target page closed view', () => {
