+++ src/tests/unit/tests/common/components/cards/get-labelled-string-property-card-row.test.tsx
