+++ src/tests/unit/tests/common/components/cards/instance-details-group.test.tsx
