+++ src/tests/unit/tests/common/components/cards/instance-details.test.tsx
