+++ src/tests/unit/tests/common/components/cards/result-section-content.test.tsx
