+++ src/tests/unit/tests/common/components/cards/result-section-title.test.tsx
