+++ src/tests/unit/tests/common/components/cards/result-section.test.tsx
