+++ src/tests/unit/tests/common/components/cards/sample-view-model-data.ts
